+++ test/chat/db/change_tracking_test.exs
@@ -28,18 +28,17 @@ defmodule Chat.Db.ChangeTrackingTest do
   test "should expire and respond" do
     pid = self()
 
-    # assert_raise ExUnit.AssertionError,
-    #              ~r/test should not rely on ChangeTracker expiration/,
-    #              fn ->
-    Tracking.new()
-    |> Tracking.add_promise(
-      :to_be_found,
-      {fn -> send(pid, :got_it) end, fn -> send(pid, :exp) end},
-      1
-    )
-    |> Tracking.extract_expired(50)
-
-    # end
+    assert_raise ExUnit.AssertionError,
+                 ~r/test should not rely on ChangeTracker expiration/,
+                 fn ->
+                   Tracking.new()
+                   |> Tracking.add_promise(
+                     :to_be_found,
+                     {fn -> send(pid, :got_it) end, fn -> send(pid, :exp) end},
+                     1
+                   )
+                   |> Tracking.extract_expired(50)
+                 end
 
     assert_receive :exp, 100
   end